Reapply "objtool/x86: Separate arch-specific and generic parts"
This reverts commit 8d38cb67607c20878b01f442dbf111cbb3f45838.
